refactor: extract settings menu from settings reducer
The settings reducers is persistent, but we don't want state data
relating to the settings menu to persist. So extract things related to
the settings menu into a dedicated reducer, leaving the settings
reducer to deal exclusively with persistent settings.
